pyproject.toml
[tool.poetry]
name = "expense-tracker-bot"
version = "0.1.0"
description = "Telegram bot for expense tracking with natural language processing"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
python-telegram-bot = "^20.8"
python-dotenv = "^1.0.0"
openai = "^1.12.0"
pandas = "^2.2.0"
matplotlib = "^3.8.2"
SQLAlchemy = "^2.0.27"
python-dateutil = "^2.8.2"
fastapi = "^0.110.0"
uvicorn = "^0.27.1"
python-multipart = "^0.0.9"
jinja2 = "^3.1.3"
aiofiles = "^23.2.1"
itsdangerous = "^2.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api" 